There's a MuLan trainer I missed...
So I think the last bit is either pre-tokenising the caption texts, or finding out if it's possible to just slap the raw texts in there. 

Conceptually would be easier for testing for now, just to put raw texts and raw wavs in. Then have a look at optimisations by thinking about pre-tokenisation of texts and saving wavs as the VQ representation, instead of actual whole wavs.